Move drawing components to components folder
