+++ src/components/LineDrawer.tsx
@@ -1,4 +1,4 @@
-// Area drawing component derived from LineDrawer
+// Line drawing component extracted from CesiumViewer
 import { useState, useRef, useEffect, useCallback } from 'react'
 import {
   Viewer,
@@ -8,27 +8,20 @@ import {
   Color,
   ColorMaterialProperty,
   ConstantProperty,
+  Property,
   ConstantPositionProperty,
   Cartesian3,
-  Cartesian2,
   Entity,
-  Property,
   HeightReference,
-  EllipsoidTangentPlane,
-  PolygonHierarchy,
-  LabelStyle,
-  VerticalOrigin,
-  Cartographic,
-  LabelGraphics,
 } from 'cesium'
-import AxisHelper from './components/axis-helper'
-import type { AxisHelperProps } from './components/axis-helper'
+import AxisHelper from './axis-helper'
+import type { AxisHelperProps } from './axis-helper'
 
-interface AreaProps {
+interface LineDrawerProps {
   viewer: Viewer | null
 }
 
-const Area = ({ viewer }: AreaProps) => {
+const LineDrawer = ({ viewer }: LineDrawerProps) => {
   const drawHandlerRef = useRef<ScreenSpaceEventHandler | null>(null)
   const selectionHandlerRef = useRef<ScreenSpaceEventHandler | null>(null)
   const startPositionRef = useRef<Cartesian3 | null>(null)
@@ -37,171 +30,16 @@ const Area = ({ viewer }: AreaProps) => {
   const mousePositionRef = useRef<Cartesian3 | null>(null)
   const selectedLineRef = useRef<Entity | null>(null)
   const selectedAnchorRef = useRef<Entity | null>(null)
-  const selectedAreaRef = useRef<Entity | null>(null)
   const [axisProps, setAxisProps] = useState<AxisHelperProps | null>(null)
-  const axisAreaRef = useRef<Entity | null>(null)
-  const movedPositionsRef = useRef<Cartesian3[] | null>(null)
-  const hierarchyCallbackRef = useRef<CallbackProperty | null>(null)
   const anchorsRef = useRef<Entity[]>([])
-  const linesRef = useRef<Entity[]>([])
-  const firstAnchorRef = useRef<Entity | null>(null)
-const polygonPositionsRef = useRef<Cartesian3[]>([])
-const [isAreaMode, setIsAreaMode] = useState(false)
-
-const computeAreaAndCentroid = useCallback(
-  (
-    positions: Cartesian3[],
-  ): { area: number; centroid: Cartesian3 } | null => {
-    if (!viewer || positions.length < 3) {
-      return null
-    }
-    const plane = EllipsoidTangentPlane.fromPoints(
-      positions,
-      viewer.scene.globe.ellipsoid,
-    )
-    const projected = plane.projectPointsOntoPlane(positions, [])
-    if (projected.length < 3) {
-      return null
-    }
-    let signedArea = 0
-    let cx = 0
-    let cy = 0
-    for (let i = 0, j = projected.length - 1; i < projected.length; j = i++) {
-      const p0 = projected[j]
-      const p1 = projected[i]
-      const f = p0.x * p1.y - p1.x * p0.y
-      signedArea += f
-      cx += (p0.x + p1.x) * f
-      cy += (p0.y + p1.y) * f
-    }
-    signedArea *= 0.5
-    if (signedArea === 0) {
-      return null
-    }
-    const area = Math.abs(signedArea)
-    cx /= 6 * signedArea
-    cy /= 6 * signedArea
-    const centroid2D = new Cartesian2(cx, cy)
-    const centroid = plane.projectPointOntoEllipsoid(
-      centroid2D,
-      new Cartesian3(),
-    )
-    return { area, centroid }
-  },
-  [viewer],
-)
-
-const computeSurfaceAreaAndCentroid = useCallback(
-  (
-    positions: Cartesian3[],
-  ): { area: number; centroid: Cartesian3 } | null => {
-    if (positions.length < 3) {
-      return null
-    }
-    let area = 0
-    const centroid = new Cartesian3(0, 0, 0)
-    const base = positions[0]
-    for (let i = 1; i < positions.length - 1; i++) {
-      const b = positions[i]
-      const c = positions[i + 1]
-      const ab = Cartesian3.subtract(b, base, new Cartesian3())
-      const ac = Cartesian3.subtract(c, base, new Cartesian3())
-      const cross = Cartesian3.cross(ab, ac, new Cartesian3())
-      const triArea = Cartesian3.magnitude(cross) * 0.5
-      area += triArea
-      const triCentroid = Cartesian3.multiplyByScalar(
-        Cartesian3.add(
-          base,
-          Cartesian3.add(b, c, new Cartesian3()),
-          new Cartesian3(),
-        ),
-        1 / 3,
-        new Cartesian3(),
-      )
-      Cartesian3.multiplyByScalar(triCentroid, triArea, triCentroid)
-      Cartesian3.add(centroid, triCentroid, centroid)
-    }
-    Cartesian3.divideByScalar(centroid, area, centroid)
-    return { area, centroid }
-  },
-  [],
-)
-
-const computeAreaWithTerrain = useCallback(
-  async (
-    positions: Cartesian3[],
-  ): Promise<{ area: number; centroid: Cartesian3 } | null> => {
-    if (!viewer || positions.length < 3) {
-      return null
-    }
-    const cartographics = positions.map((p) =>
-      Cartographic.fromCartesian(p),
-    )
-    try {
-      const sampled = await viewer.scene.sampleHeightMostDetailed(cartographics)
-      const withHeights = sampled.map((c) =>
-        Cartesian3.fromRadians(c.longitude, c.latitude, c.height),
-      )
-      return computeSurfaceAreaAndCentroid(withHeights)
-    } catch {
-      return computeSurfaceAreaAndCentroid(positions)
-    }
-  },
-  [viewer, computeSurfaceAreaAndCentroid],
-)
-
-
-
-
-
-  const removeAxisHelper = useCallback(async () => {
-    if (!viewer) {
-      return
-    }
-    setAxisProps(null)
-    if (axisAreaRef.current && movedPositionsRef.current) {
-      const area = axisAreaRef.current
-      const positions = movedPositionsRef.current
-      axisAreaRef.current = null
-      movedPositionsRef.current = null
-      hierarchyCallbackRef.current = null
-
-      area.polygon!.hierarchy = new ConstantProperty(
-        new PolygonHierarchy(positions),
-      )
-      ;(area as Entity & { positions?: Cartesian3[] }).positions = positions
-      const result = await computeAreaWithTerrain(positions)
-      if (result) {
-        area.position = new ConstantPositionProperty(result.centroid)
-        if (area.label) {
-          area.label.text = new ConstantProperty(
-            `${Math.round(result.area)} m²`,
-          )
-        } else {
-          area.label = new LabelGraphics({
-            text: new ConstantProperty(`${Math.round(result.area)} m²`),
-            fillColor: new ConstantProperty(Color.BLACK),
-            style: new ConstantProperty(LabelStyle.FILL),
-            showBackground: new ConstantProperty(true),
-            backgroundColor: new ConstantProperty(
-              Color.WHITE.withAlpha(0.5),
-            ),
-            verticalOrigin: new ConstantProperty(VerticalOrigin.CENTER),
-            heightReference: new ConstantProperty(
-              HeightReference.CLAMP_TO_GROUND,
-            ),
-          })
-        }
-      }
-    }
-  }, [viewer, computeAreaWithTerrain])
+  const [isLineMode, setIsLineMode] = useState(false)
 
   const showLineAxis = useCallback(
     (line: Entity) => {
       if (!viewer) {
         return
       }
-      removeAxisHelper()
+      setAxisProps(null)
       const update = (translation: Cartesian3) => {
         const time = viewer.clock.currentTime
         const positions =
@@ -241,78 +79,7 @@ const computeAreaWithTerrain = useCallback(
         onTranslate: update,
       })
     },
-    [viewer, removeAxisHelper],
-  )
-
-  const showAxisHelper = useCallback(
-    (area: Entity) => {
-      if (!viewer) {
-        return
-      }
-      removeAxisHelper()
-      movedPositionsRef.current = (
-        (area as Entity & { positions?: Cartesian3[] }).positions || []
-      ).map((p) => Cartesian3.clone(p))
-      hierarchyCallbackRef.current = new CallbackProperty(() => {
-        return new PolygonHierarchy(movedPositionsRef.current || [])
-      }, false)
-      area.polygon!.hierarchy = hierarchyCallbackRef.current
-      axisAreaRef.current = area
-
-      const update = (translation: Cartesian3) => {
-        const pos = area.position?.getValue(viewer.clock.currentTime)
-        if (!pos || !movedPositionsRef.current) return
-        const newPos = Cartesian3.add(pos, translation, new Cartesian3())
-        area.position = new ConstantPositionProperty(newPos)
-        movedPositionsRef.current = movedPositionsRef.current.map((p) =>
-          Cartesian3.add(p, translation, new Cartesian3()),
-        )
-        ;(area as Entity & { positions?: Cartesian3[] }).positions =
-          movedPositionsRef.current
-        const result = computeAreaAndCentroid(movedPositionsRef.current)
-        if (result) {
-          area.position = new ConstantPositionProperty(result.centroid)
-          if (area.label) {
-            area.label.text = new ConstantProperty(`${Math.round(result.area)} m²`)
-          }
-        }
-      }
-
-      setAxisProps({
-        viewer,
-        enableZ: false,
-        getPosition: () =>
-          area.position?.getValue(viewer.clock.currentTime) || null,
-        onTranslate: update,
-      })
-    },
-    [viewer, removeAxisHelper, computeAreaAndCentroid],
-  )
-
-  const highlightArea = useCallback(
-    (area: Entity) => {
-      if (area.polygon) {
-        area.polygon.material = new ColorMaterialProperty(
-          Color.RED.withAlpha(0.5),
-        )
-        area.polygon.outlineColor = new ConstantProperty(Color.RED)
-      }
-      showAxisHelper(area)
-    },
-    [showAxisHelper],
-  )
-
-  const unhighlightArea = useCallback(
-    (area: Entity) => {
-      if (area.polygon) {
-        area.polygon.material = new ColorMaterialProperty(
-          Color.YELLOW.withAlpha(0.5),
-        )
-        area.polygon.outlineColor = new ConstantProperty(Color.YELLOW)
-      }
-      removeAxisHelper()
-    },
-    [removeAxisHelper],
+    [viewer],
   )
 
   const highlightLine = useCallback(
@@ -332,9 +99,9 @@ const computeAreaWithTerrain = useCallback(
         line.polyline.material = new ColorMaterialProperty(Color.YELLOW)
         line.polyline.width = new ConstantProperty(2)
       }
-      removeAxisHelper()
+      setAxisProps(null)
     },
-    [removeAxisHelper],
+    [],
   )
 
   const highlightAnchor = useCallback((anchor: Entity) => {
@@ -351,14 +118,12 @@ const computeAreaWithTerrain = useCallback(
     }
   }, [])
 
-
   const removeLine = useCallback(
     (line: Entity) => {
     if (!viewer) {
       return
     }
     viewer.entities.remove(line)
-    linesRef.current = linesRef.current.filter((l) => l !== line)
     const lineWithAnchors = line as Entity & { anchors?: [Entity, Entity] }
     if (lineWithAnchors.anchors) {
       for (const anchor of lineWithAnchors.anchors) {
@@ -394,19 +159,6 @@ const computeAreaWithTerrain = useCallback(
     [removeLine, viewer],
   )
 
-  const removeArea = useCallback(
-    (area: Entity) => {
-      if (!viewer) {
-        return
-      }
-      viewer.entities.remove(area)
-      if (selectedAreaRef.current === area) {
-        removeAxisHelper()
-      }
-    },
-    [viewer, removeAxisHelper],
-  )
-
   const addAnchor = (position: Cartesian3) => {
     if (!viewer) {
       return null
@@ -433,7 +185,7 @@ const computeAreaWithTerrain = useCallback(
     return anchor
   }
 
-  const startAreaMode = () => {
+  const startLineMode = () => {
     if (!viewer) {
       return
     }
@@ -441,7 +193,7 @@ const computeAreaWithTerrain = useCallback(
     if (drawHandlerRef.current) {
       drawHandlerRef.current.destroy()
       drawHandlerRef.current = null
-      setIsAreaMode(false)
+      setIsLineMode(false)
       if (drawingLineRef.current) {
         viewer.entities.remove(drawingLineRef.current)
         drawingLineRef.current = null
@@ -463,7 +215,7 @@ const computeAreaWithTerrain = useCallback(
 
     const handler = new ScreenSpaceEventHandler(viewer.scene.canvas)
     drawHandlerRef.current = handler
-    setIsAreaMode(true)
+    setIsLineMode(true)
 
     let isDrawing = false
     let longPressTimeout: number | null = null
@@ -502,8 +254,6 @@ const computeAreaWithTerrain = useCallback(
         startPositionRef.current = position
         mousePositionRef.current = position
         startAnchorRef.current = addAnchor(position)!
-        firstAnchorRef.current = startAnchorRef.current
-        polygonPositionsRef.current = [position]
         const dynamicPositions = new CallbackProperty(() => {
           if (!startPositionRef.current || !mousePositionRef.current) {
             return []
@@ -533,46 +283,6 @@ const computeAreaWithTerrain = useCallback(
         ]
         ;(startAnchorRef.current! as Entity & { connectedLines: Set<Entity> }).connectedLines.add(line)
         ;(endAnchor as Entity & { connectedLines: Set<Entity> }).connectedLines.add(line)
-        linesRef.current.push(line)
-        polygonPositionsRef.current.push(position)
-        if (
-          endAnchor === firstAnchorRef.current &&
-          polygonPositionsRef.current.length >= 3
-        ) {
-          const polyPositions = [...polygonPositionsRef.current]
-          const result = computeAreaAndCentroid(polyPositions)
-          const areaEntity = viewer.entities.add({
-            position: result?.centroid,
-            polygon: {
-              hierarchy: new PolygonHierarchy(polyPositions),
-              material: new ColorMaterialProperty(Color.YELLOW.withAlpha(0.5)),
-              outline: true,
-              outlineColor: Color.YELLOW,
-              heightReference: HeightReference.CLAMP_TO_GROUND,
-            },
-            label: result
-              ? {
-                  text: `${Math.round(result.area)} m²`,
-                  fillColor: Color.BLACK,
-                  style: LabelStyle.FILL,
-                  showBackground: true,
-                  backgroundColor: Color.WHITE.withAlpha(0.5),
-                  verticalOrigin: VerticalOrigin.CENTER,
-                  heightReference: HeightReference.CLAMP_TO_GROUND,
-                }
-              : undefined,
-          })
-          ;(areaEntity as Entity & { isArea: boolean; positions: Cartesian3[] }).isArea = true
-          ;(areaEntity as Entity & { positions: Cartesian3[] }).positions = polyPositions
-          for (const l of linesRef.current) {
-            removeLine(l)
-          }
-          linesRef.current = []
-          firstAnchorRef.current = null
-          polygonPositionsRef.current = []
-          finishDrawing()
-          return
-        }
         startPositionRef.current = position
         startAnchorRef.current = endAnchor
         mousePositionRef.current = position
@@ -629,8 +339,6 @@ const computeAreaWithTerrain = useCallback(
       }
       startPositionRef.current = null
       mousePositionRef.current = null
-      firstAnchorRef.current = null
-      polygonPositionsRef.current = []
       isDrawing = false
     }
 
@@ -665,7 +373,6 @@ const computeAreaWithTerrain = useCallback(
           const entity = picked.id as Entity & {
             isLine?: boolean
             isAnchor?: boolean
-            isArea?: boolean
           }
           if (entity.isLine) {
             if (selectedLineRef.current && selectedLineRef.current !== entity) {
@@ -674,10 +381,6 @@ const computeAreaWithTerrain = useCallback(
             if (selectedAnchorRef.current) {
               unhighlightAnchor(selectedAnchorRef.current)
               selectedAnchorRef.current = null
-            }
-            if (selectedAreaRef.current) {
-              unhighlightArea(selectedAreaRef.current)
-              selectedAreaRef.current = null
             }
             selectedLineRef.current = entity
             highlightLine(entity)
@@ -694,31 +397,8 @@ const computeAreaWithTerrain = useCallback(
               unhighlightLine(selectedLineRef.current)
               selectedLineRef.current = null
             }
-            if (selectedAreaRef.current) {
-              unhighlightArea(selectedAreaRef.current)
-              selectedAreaRef.current = null
-            }
             selectedAnchorRef.current = entity
             highlightAnchor(entity)
-            return
-          }
-          if (entity.isArea) {
-            if (
-              selectedAreaRef.current &&
-              selectedAreaRef.current !== entity
-            ) {
-              unhighlightArea(selectedAreaRef.current)
-            }
-            if (selectedLineRef.current) {
-              unhighlightLine(selectedLineRef.current)
-              selectedLineRef.current = null
-            }
-            if (selectedAnchorRef.current) {
-              unhighlightAnchor(selectedAnchorRef.current)
-              selectedAnchorRef.current = null
-            }
-            selectedAreaRef.current = entity
-            highlightArea(entity)
             return
           }
         }
@@ -730,10 +410,6 @@ const computeAreaWithTerrain = useCallback(
           unhighlightAnchor(selectedAnchorRef.current)
           selectedAnchorRef.current = null
         }
-        if (selectedAreaRef.current) {
-          unhighlightArea(selectedAreaRef.current)
-          selectedAreaRef.current = null
-        }
       },
       ScreenSpaceEventType.LEFT_CLICK,
     )
@@ -744,8 +420,6 @@ const computeAreaWithTerrain = useCallback(
     }
   }, [
     viewer,
-    highlightArea,
-    unhighlightArea,
     highlightLine,
     unhighlightLine,
     highlightAnchor,
@@ -757,7 +431,7 @@ const computeAreaWithTerrain = useCallback(
       if (event.key === 'Escape' && drawHandlerRef.current) {
         drawHandlerRef.current.destroy()
         drawHandlerRef.current = null
-        setIsAreaMode(false)
+        setIsLineMode(false)
         if (drawingLineRef.current) {
           viewer?.entities.remove(drawingLineRef.current)
           drawingLineRef.current = null
@@ -782,9 +456,6 @@ const computeAreaWithTerrain = useCallback(
         } else if (selectedAnchorRef.current) {
           removeAnchor(selectedAnchorRef.current)
           selectedAnchorRef.current = null
-        } else if (selectedAreaRef.current) {
-          removeArea(selectedAreaRef.current)
-          selectedAreaRef.current = null
         }
       }
     }
@@ -792,27 +463,19 @@ const computeAreaWithTerrain = useCallback(
     return () => {
       window.removeEventListener('keydown', handleKeyDown)
     }
-  }, [
-    isAreaMode,
-    removeLine,
-    removeAnchor,
-    removeArea,
-    viewer,
-    highlightArea,
-    unhighlightArea,
-  ])
+  }, [isLineMode, removeLine, removeAnchor, viewer])
 
   return (
     <>
       {axisProps && <AxisHelper {...axisProps} />}
       <button
-        onClick={startAreaMode}
-        style={{ border: isAreaMode ? '2px solid yellow' : '1px solid gray' }}
+        onClick={startLineMode}
+        style={{ border: isLineMode ? '2px solid yellow' : '1px solid gray' }}
       >
-        Area
+        Line
       </button>
     </>
   )
 }
 
-export default Area
+export default LineDrawer
