+++ src/components/ToolsPanel.tsx
